refactor: group topbar, mainview and their components
