+++ frontend/src/components/MainView/ColorToggle/index.tsx
