+++ frontend/src/components/MainView/ColorToggle/style.css
