+++ frontend/src/components/MainView/Legend/style.css
