+++ frontend/src/components/MainView/ZoomButtons/index.tsx
