+++ frontend/src/components/MainView/ZoomButtons/style.css
