+++ frontend/src/components/MainView/style.css
