+++ frontend/src/components/TopBar/TopBarMenu/index.tsx
@@ -1,6 +1,6 @@
 import { LogOut, X } from "lucide-react";
 import { useEffect, useRef } from "react";
-import { useAuth } from "../hooks/useAuth";
+import { useAuth } from "../../../hooks/useAuth";
 import "./TopBarMenu.css";
 
 interface TopBarMenuProps {
