+++ frontend/tests/components/mainView.test.tsx
@@ -3,7 +3,7 @@ import { render, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import type { ReactElement } from "react";
 import { describe, expect, it, vi } from "vitest";
-import MainView from "../../src/components/MainView.tsx";
+import MainView from "../../src/components/MainView/index.tsx";
 import { RoomSelectionProvider } from "../../src/components/RoomSelectionProvider.tsx";
 import {
   DEFAULT_SCALE,
